Added safe mode for plugin error capturing reorganized source files structure
