--- js-framework/src/plugin-manager.tsx
@@ -1,397 +1,0 @@
-import BetterNCM from "./betterncm-api";
-import { loadedPlugins, NCMPlugin } from "./loader";
-
-const configKey = "config.betterncm.manager.open";
-
-export async function initPluginManager() {
-	// 准备设置页面和访问按钮
-	const settingsView = document.createElement("section");
-	const mainPageView: HTMLElement = (await BetterNCM.utils.waitForElement(
-		"section.g-mn",
-	))!!;
-	const settingsButton = (await BetterNCM.utils.waitForElement(
-		'a[href="#/m/setting/"]',
-	))!! as HTMLAnchorElement;
-	const lyricButton = (await BetterNCM.utils.waitForElement(
-		"div.cover.u-cover.u-cover-sm > a > span",
-	))!! as HTMLAnchorElement;
-	const betterNCMSettingsButton = settingsButton.cloneNode(
-		true,
-	) as HTMLAnchorElement;
-	betterNCMSettingsButton.href = "javascript:void(0)";
-	betterNCMSettingsButton.title = "BetterNCM";
-	betterNCMSettingsButton.innerHTML = `<svg style='transform: scale(0.8);'><use xlink:href="orpheus://orpheus/style/res/svg/topbar.sp.svg#logo_white"></use></svg>`;
-	mainPageView.parentElement!!.insertBefore(
-		settingsView,
-		mainPageView.nextElementSibling,
-	);
-	settingsButton.parentElement!!.insertBefore(
-		betterNCMSettingsButton,
-		settingsButton.nextElementSibling,
-	);
-	ReactDOM.render(<PluginManager />, settingsView);
-
-	settingsView.classList.add("better-ncm-manager");
-
-	function showSettings() {
-		// 有插件似乎会替换主页元素，导致我们的设置页面无法显示，需要进行检查
-		if (settingsView.parentElement !== mainPageView.parentElement) {
-			mainPageView.parentElement!!.insertBefore(
-				settingsView,
-				mainPageView.nextElementSibling,
-			);
-		}
-		settingsView.classList.add("ncmm-show");
-		// 有些主题插件会给我们主页上 !important 优先级修饰符
-		// 所以得这样硬碰硬
-		mainPageView.setAttribute("style", "display: none !important;");
-	}
-
-	function hideSettings() {
-		settingsView.classList.remove("ncmm-show");
-		mainPageView.removeAttribute("style");
-	}
-
-	lyricButton.addEventListener("click", hideSettings);
-	settingsButton.addEventListener("click", hideSettings);
-	betterNCMSettingsButton.addEventListener("click", () => {
-		if (settingsView.classList.contains("ncmm-show")) {
-			hideSettings();
-		} else {
-			showSettings();
-		}
-	});
-
-	// 如果外部页面变更（点击了其它按钮跳转）则关闭设置页面
-	window.addEventListener("hashchange", hideSettings);
-	new MutationObserver((rs) => {
-		for (const r of rs) {
-			if (r.attributeName === "style") {
-				settingsView.style.left = mainPageView.style.left;
-			}
-		}
-	}).observe(mainPageView, {
-		attributes: true,
-	});
-
-	// g-mn-set
-}
-
-const ProgressRing: React.FC<{
-	size?: string;
-}> = (props) => {
-	return (
-		<span
-			className="bncm-spinner"
-			style={{
-				width: props.size || "16px",
-				height: props.size || "16px",
-			}}
-		>
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-			<div />
-		</span>
-	);
-};
-
-export let onPluginLoaded = (_: typeof loadedPlugins) => {};
-
-const Button: React.FC<
-	React.PropsWithChildren<React.HTMLAttributes<HTMLAnchorElement>>
-> = (props) => {
-	const { children, className, ...other } = props;
-	return (
-		<a className={`u-ibtn5 u-ibtnsz8 ${className || ""}`} {...other}>
-			{children}
-		</a>
-	);
-};
-
-const PluginManager: React.FC = () => {
-	interface ReleaseVersion {
-		version: string;
-		supports: string[];
-		file: string;
-		changelog: string;
-	}
-
-	interface OnlineVersionInfo {
-		versions: ReleaseVersion[];
-	}
-
-	const [currentVersion, setCurrentVersion] = React.useState("");
-	const [selectedPlugin, setSelectedPlugin] = React.useState<NCMPlugin | null>(
-		null,
-	);
-	const [latestVersion, setLatestVersion] =
-		React.useState<ReleaseVersion | null>(null);
-	const [updateButtonColor, setUpdateButtonColor] =
-		React.useState("transparent"); // #F004 #0F04
-	const headerRef = React.useRef<HTMLElement | null>(null);
-	const pluginConfigRef = React.useRef<HTMLDivElement | null>(null);
-	const [borderColor, setBorderColor] = React.useState("");
-	const [loadedPluginsList, setLoadedPlugins] = React.useState<string[]>([]);
-
-	React.useEffect(() => {
-		// 获取边框颜色
-		if (headerRef.current) {
-			setBorderColor(
-				window
-					.getComputedStyle(headerRef.current, null)
-					.getPropertyValue("border-bottom-color"),
-			);
-		}
-	}, [headerRef.current]);
-
-	React.useEffect(() => {
-		const myDomElement =
-			(selectedPlugin?.injects
-				.map((v) => v._getConfigElement())
-				.filter((v) => v !== null) as HTMLElement[] | null) || [];
-
-		if (myDomElement.length === 0) {
-			const tipElement = document.createElement("div");
-			tipElement.innerText = "该插件没有可用的设置选项";
-			myDomElement.push(tipElement);
-		}
-
-		pluginConfigRef.current?.replaceChildren(...myDomElement);
-	}, [selectedPlugin]);
-
-	React.useEffect(() => {
-		function sortFunc(key1: string, key2: string) {
-			const getSortValue = (key: string) => {
-				const loadPlugin = loadedPlugins[key];
-				const value = loadPlugin.haveConfigElement() ? 1 : 0;
-
-				// 将插件商店排到最前面
-				if (loadPlugin.manifest.name.startsWith("PluginMarket"))
-					return Number.MAX_SAFE_INTEGER;
-
-				return value;
-			};
-			return getSortValue(key2) - getSortValue(key1);
-		}
-		setLoadedPlugins(Object.keys(loadedPlugins).sort(sortFunc));
-		onPluginLoaded = (loadedPlugins) => {
-			console.log("插件加载完成！");
-			setLoadedPlugins(Object.keys(loadedPlugins).sort(sortFunc));
-		};
-		(async () => {
-			if (!latestVersion) {
-				const betterNCMVersion = await BetterNCM.app.getBetterNCMVersion();
-				setCurrentVersion(betterNCMVersion);
-				const currentNCMVersion = BetterNCM.ncm.getNCMVersion();
-
-				const online: OnlineVersionInfo = await (
-					await fetch(
-						"https://gitee.com/microblock/better-ncm-v2-data/raw/master/betterncm/betterncm.json",
-					)
-				).json();
-				const onlineSuitableVersions = online.versions.filter((v) =>
-					v.supports.includes(currentNCMVersion),
-				);
-				if (onlineSuitableVersions.length === 0) {
-					setUpdateButtonColor("#F004");
-					setLatestVersion({
-						version: "",
-						supports: [],
-						file: "",
-						changelog: "",
-					});
-				} else {
-					const latestVersion = onlineSuitableVersions[0];
-					if (latestVersion.version !== betterNCMVersion) {
-						setUpdateButtonColor("#0F04");
-					}
-					setLatestVersion(latestVersion);
-				}
-			}
-		})();
-	}, [latestVersion]);
-
-	const onUpdateButtonClicked = React.useMemo(
-		() => async () => {
-			if (latestVersion && latestVersion.version !== currentVersion) {
-				const ncmpath = await BetterNCM.app.getNCMPath();
-				const datapath = await BetterNCM.app.getDataPath();
-				const dllpath = `${datapath}\\betterncm.dll`;
-				if (await BetterNCM.fs.exists("./betterncm.dll"))
-					await BetterNCM.fs.remove("./betterncm.dll");
-
-				await BetterNCM.fs.writeFile(
-					"./betterncm.dll",
-					await (await fetch(latestVersion?.file)).blob(),
-				);
-
-				if (!ncmpath.toLowerCase().includes("system")) {
-					BetterNCM.app.exec(
-						`cmd /c @echo off & echo BetterNCM Updating... & cd /d C:/ & cd C:/ & cd /d ${ncmpath[0]}:/ & cd "${ncmpath}" & taskkill /f /im cloudmusic.exe>nul & taskkill /f /im cloudmusicn.exe>nul & ping 127.0.0.1>nul & del msimg32.dll & move "${dllpath}" .\\msimg32.dll & start cloudmusic.exe`,
-						true,
-					);
-				}
-			} else if (latestVersion) {
-				// 重新检测新版本
-				setLatestVersion(null);
-			}
-		},
-		[],
-	);
-
-	return (
-		<div className="bncm-mgr">
-			<div>
-				<section className="bncm-mgr-header" ref={headerRef}>
-					<img
-						src="https://s1.ax1x.com/2022/08/11/vGlJN8.png"
-						alt=""
-						style={{
-							height: "64px",
-						}}
-					/>
-					<div>
-						<h1>
-							BetterNCM{" "}
-							<span style={{ fontSize: "smaller", opacity: "0.8" }}>0.2.5</span>
-						</h1>
-						<div className="bncm-mgr-btns">
-							<Button
-								onClick={async () => {
-									BetterNCM.app.exec(
-										`explorer "${(await BetterNCM.app.getDataPath()).replace(
-											/\//g,
-											"\\",
-										)}"`,
-										false,
-										true,
-									);
-								}}
-							>
-								打开插件文件夹
-							</Button>
-							<Button
-								onClick={() => {
-									BetterNCM.app.showConsole();
-								}}
-							>
-								打开控制台
-							</Button>
-							<Button
-								onClick={async () => {
-									await BetterNCM.app.reloadPlugins();
-									BetterNCM.reload();
-								}}
-							>
-								重载插件
-							</Button>
-							<Button
-								style={{
-									display: "flex",
-									alignItems: "center",
-									background: updateButtonColor,
-								}}
-								onClick={onUpdateButtonClicked}
-							>
-								{latestVersion === null ? (
-									<>
-										<ProgressRing />
-										检查更新中
-									</>
-								) : latestVersion.version === currentVersion ? (
-									<>已是最新版本</>
-								) : latestVersion.version.length === 0 ? (
-									<>版本不兼容</>
-								) : (
-									<>点击更新到 {latestVersion.version}</>
-								)}
-							</Button>
-						</div>
-					</div>
-					<div className="m-tool">
-						<a
-							className="itm"
-							// rome-ignore lint/a11y/useValidAnchor: <explanation>
-							onClick={() =>
-								BetterNCM.ncm.openUrl("https://github.com/MicroCBer/BetterNCM")
-							}
-							style={{
-								width: "32px",
-								height: "32px",
-							}}
-						>
-							<svg width="32" height="32" viewBox="0 0 16 16" version="1.1">
-								<path
-									fillRule="evenodd"
-									d="M8 0C3.58 0 0 3.58 0 8c0 3.54 2.29 6.53 5.47 7.59.4.07.55-.17.55-.38 0-.19-.01-.82-.01-1.49-2.01.37-2.53-.49-2.69-.94-.09-.23-.48-.94-.82-1.13-.28-.15-.68-.52-.01-.53.63-.01 1.08.58 1.23.82.72 1.21 1.87.87 2.33.66.07-.52.28-.87.51-1.07-1.78-.2-3.64-.89-3.64-3.95 0-.87.31-1.59.82-2.15-.08-.2-.36-1.02.08-2.12 0 0 .67-.21 2.2.82.64-.18 1.32-.27 2-.27.68 0 1.36.09 2 .27 1.53-1.04 2.2-.82 2.2-.82.44 1.1.16 1.92.08 2.12.51.56.82 1.27.82 2.15 0 3.07-1.87 3.75-3.65 3.95.29.25.54.73.54 1.48 0 1.07-.01 1.93-.01 2.2 0 .21.15.46.55.38A8.013 8.013 0 0016 8c0-4.42-3.58-8-8-8z"
-								/>
-							</svg>
-						</a>
-					</div>
-				</section>
-				<section
-					style={{
-						display: "flex",
-						flexDirection: "row",
-						flex: "1",
-						marginBottom: "0",
-					}}
-				>
-					<div
-						className="v-scroll loaded-plugins-list"
-						style={{
-							borderRight: `1px solid ${borderColor}`,
-						}}
-					>
-						<div>
-							<div>
-								{loadedPluginsList.map((key) => {
-									const loadPlugin = loadedPlugins[key];
-									const haveConfig = loadPlugin.haveConfigElement();
-									return (
-										// rome-ignore lint/a11y/useKeyWithClickEvents: <explanation>
-										<div
-											className={
-												haveConfig
-													? selectedPlugin?.manifest.name === key
-														? "plugin-btn selected"
-														: "plugin-btn"
-													: "plugin-btn-disabled"
-											}
-											onClick={() => {
-												if (haveConfig) setSelectedPlugin(loadPlugin);
-											}}
-										>
-											{loadPlugin.manifest.name}
-										</div>
-									);
-								})}
-							</div>
-						</div>
-					</div>
-					<div className="v-scroll">
-						<div>
-							<div
-								style={{
-									overflowY: "scroll",
-									overflowX: "hidden",
-									padding: "16px",
-								}}
-								ref={pluginConfigRef}
-							/>
-						</div>
-					</div>
-				</section>
-			</div>
-		</div>
-	);
-};
